requirements.txt
rich
requests
scapy
capstone
pyelftools
bitstring
python-dotenv
